Fixed a bunch of key spacing issues
I put in some more work on the case, then realized that if the keys are too close on the pcb, the keycaps wont fit (i was worried about larger keys like shift, and how far they should be spaced from the letter keys). i did some googling, and realized that there is an actual grid system for this, and I had completely ignored this when placing keys on my PCB. I had to go back over every key and measure the correct distances, then adjust the switch position accordingly. It was a lot of tedious work, but it payed off. It would have sucked to order the PCB and have none of the keycaps fit.